Added method : - (void)rotateCardAtIndex: - (void)openCards:
Was created general structure for describing gamer rates: raise, fold, check...
added new small methods.

+++ Poker/Gamer.h
@@ -21,7 +21,7 @@
 @property(nonatomic,strong)NSString *ipAddress;
 @property(nonatomic)int port;
 
-@property(nonatomic)int rate;
+@property(nonatomic)long rate;
 @property(nonatomic)BOOL isGamed;
